Add the helm computer
Add all the required files, with what I would assume to be placeholder names as I have no idea what I should call them otherwise. The whole point of this branch is to add a new type of shuttle computer thats only able to be built, and not taken apart. That way it can be used in a potential ship map without the risk of shittle builders uninstalling the board